+++ SF3Lib/Models/Structs/CHR/MultipleAnimationOffsetsRow.cs
@@ -53,10 +53,10 @@ namespace SF3.Models.Structs.CHR {
         public int SpriteID { get; }
 
         public uint this[int index] {
-            get => (index >= 0 && index < 16) ? (uint) Data.GetInt32(Address + index * 0x04) : throw new ArgumentOutOfRangeException(nameof(index));
+            get => (index >= 0 && index < 16) ? Data.GetUInt32(Address + index * 0x04) : throw new ArgumentOutOfRangeException(nameof(index));
             set {
                 if (index >= 0 && index < 16)
-                    Data.SetInt32(Address + index * 0x04, (int) value);
+                    Data.SetUInt32(Address + index * 0x04, value);
                 else
                     throw new ArgumentOutOfRangeException(nameof(index));
             }
@@ -65,7 +65,7 @@ namespace SF3.Models.Structs.CHR {
         public IEnumerable<uint> GetOffsets() {
             var offsets = new uint[16];
             for (int i = 0; i < offsets.Length; i++)
-                offsets[i] = (uint) Data.GetInt32(Address + i * 0x04);
+                offsets[i] = Data.GetUInt32(Address + i * 0x04);
             return offsets;
         }
 
@@ -73,21 +73,21 @@ namespace SF3.Models.Structs.CHR {
         IEnumerator IEnumerable.GetEnumerator() => GetOffsets().GetEnumerator();
 
         // Gigantic table of properties so it's visible to the ObjectListView (eeeeew)
-        [TableViewModelColumn(addressField: nameof(_offset01Addr), displayOrder:  0, isPointer: true)] [BulkCopy] public uint Offset01 { get => (uint) Data.GetInt32(_offset01Addr); set => Data.SetInt32(_offset01Addr, (int) value); }
-        [TableViewModelColumn(addressField: nameof(_offset02Addr), displayOrder:  1, isPointer: true)] [BulkCopy] public uint Offset02 { get => (uint) Data.GetInt32(_offset02Addr); set => Data.SetInt32(_offset02Addr, (int) value); }
-        [TableViewModelColumn(addressField: nameof(_offset03Addr), displayOrder:  2, isPointer: true)] [BulkCopy] public uint Offset03 { get => (uint) Data.GetInt32(_offset03Addr); set => Data.SetInt32(_offset03Addr, (int) value); }
-        [TableViewModelColumn(addressField: nameof(_offset04Addr), displayOrder:  3, isPointer: true)] [BulkCopy] public uint Offset04 { get => (uint) Data.GetInt32(_offset04Addr); set => Data.SetInt32(_offset04Addr, (int) value); }
-        [TableViewModelColumn(addressField: nameof(_offset05Addr), displayOrder:  4, isPointer: true)] [BulkCopy] public uint Offset05 { get => (uint) Data.GetInt32(_offset05Addr); set => Data.SetInt32(_offset05Addr, (int) value); }
-        [TableViewModelColumn(addressField: nameof(_offset06Addr), displayOrder:  5, isPointer: true)] [BulkCopy] public uint Offset06 { get => (uint) Data.GetInt32(_offset06Addr); set => Data.SetInt32(_offset06Addr, (int) value); }
-        [TableViewModelColumn(addressField: nameof(_offset07Addr), displayOrder:  6, isPointer: true)] [BulkCopy] public uint Offset07 { get => (uint) Data.GetInt32(_offset07Addr); set => Data.SetInt32(_offset07Addr, (int) value); }
-        [TableViewModelColumn(addressField: nameof(_offset08Addr), displayOrder:  7, isPointer: true)] [BulkCopy] public uint Offset08 { get => (uint) Data.GetInt32(_offset08Addr); set => Data.SetInt32(_offset08Addr, (int) value); }
-        [TableViewModelColumn(addressField: nameof(_offset09Addr), displayOrder:  8, isPointer: true)] [BulkCopy] public uint Offset09 { get => (uint) Data.GetInt32(_offset09Addr); set => Data.SetInt32(_offset09Addr, (int) value); }
-        [TableViewModelColumn(addressField: nameof(_offset10Addr), displayOrder:  9, isPointer: true)] [BulkCopy] public uint Offset10 { get => (uint) Data.GetInt32(_offset10Addr); set => Data.SetInt32(_offset10Addr, (int) value); }
-        [TableViewModelColumn(addressField: nameof(_offset11Addr), displayOrder: 10, isPointer: true)] [BulkCopy] public uint Offset11 { get => (uint) Data.GetInt32(_offset11Addr); set => Data.SetInt32(_offset11Addr, (int) value); }
-        [TableViewModelColumn(addressField: nameof(_offset12Addr), displayOrder: 11, isPointer: true)] [BulkCopy] public uint Offset12 { get => (uint) Data.GetInt32(_offset12Addr); set => Data.SetInt32(_offset12Addr, (int) value); }
-        [TableViewModelColumn(addressField: nameof(_offset13Addr), displayOrder: 12, isPointer: true)] [BulkCopy] public uint Offset13 { get => (uint) Data.GetInt32(_offset13Addr); set => Data.SetInt32(_offset13Addr, (int) value); }
-        [TableViewModelColumn(addressField: nameof(_offset14Addr), displayOrder: 13, isPointer: true)] [BulkCopy] public uint Offset14 { get => (uint) Data.GetInt32(_offset14Addr); set => Data.SetInt32(_offset14Addr, (int) value); }
-        [TableViewModelColumn(addressField: nameof(_offset15Addr), displayOrder: 14, isPointer: true)] [BulkCopy] public uint Offset15 { get => (uint) Data.GetInt32(_offset15Addr); set => Data.SetInt32(_offset15Addr, (int) value); }
-        [TableViewModelColumn(addressField: nameof(_offset16Addr), displayOrder: 15, isPointer: true)] [BulkCopy] public uint Offset16 { get => (uint) Data.GetInt32(_offset16Addr); set => Data.SetInt32(_offset16Addr, (int) value); }
+        [TableViewModelColumn(addressField: nameof(_offset01Addr), displayOrder:  0, isPointer: true)] [BulkCopy] public uint Offset01 { get => Data.GetUInt32(_offset01Addr); set => Data.SetUInt32(_offset01Addr, value); }
+        [TableViewModelColumn(addressField: nameof(_offset02Addr), displayOrder:  1, isPointer: true)] [BulkCopy] public uint Offset02 { get => Data.GetUInt32(_offset02Addr); set => Data.SetUInt32(_offset02Addr, value); }
+        [TableViewModelColumn(addressField: nameof(_offset03Addr), displayOrder:  2, isPointer: true)] [BulkCopy] public uint Offset03 { get => Data.GetUInt32(_offset03Addr); set => Data.SetUInt32(_offset03Addr, value); }
+        [TableViewModelColumn(addressField: nameof(_offset04Addr), displayOrder:  3, isPointer: true)] [BulkCopy] public uint Offset04 { get => Data.GetUInt32(_offset04Addr); set => Data.SetUInt32(_offset04Addr, value); }
+        [TableViewModelColumn(addressField: nameof(_offset05Addr), displayOrder:  4, isPointer: true)] [BulkCopy] public uint Offset05 { get => Data.GetUInt32(_offset05Addr); set => Data.SetUInt32(_offset05Addr, value); }
+        [TableViewModelColumn(addressField: nameof(_offset06Addr), displayOrder:  5, isPointer: true)] [BulkCopy] public uint Offset06 { get => Data.GetUInt32(_offset06Addr); set => Data.SetUInt32(_offset06Addr, value); }
+        [TableViewModelColumn(addressField: nameof(_offset07Addr), displayOrder:  6, isPointer: true)] [BulkCopy] public uint Offset07 { get => Data.GetUInt32(_offset07Addr); set => Data.SetUInt32(_offset07Addr, value); }
+        [TableViewModelColumn(addressField: nameof(_offset08Addr), displayOrder:  7, isPointer: true)] [BulkCopy] public uint Offset08 { get => Data.GetUInt32(_offset08Addr); set => Data.SetUInt32(_offset08Addr, value); }
+        [TableViewModelColumn(addressField: nameof(_offset09Addr), displayOrder:  8, isPointer: true)] [BulkCopy] public uint Offset09 { get => Data.GetUInt32(_offset09Addr); set => Data.SetUInt32(_offset09Addr, value); }
+        [TableViewModelColumn(addressField: nameof(_offset10Addr), displayOrder:  9, isPointer: true)] [BulkCopy] public uint Offset10 { get => Data.GetUInt32(_offset10Addr); set => Data.SetUInt32(_offset10Addr, value); }
+        [TableViewModelColumn(addressField: nameof(_offset11Addr), displayOrder: 10, isPointer: true)] [BulkCopy] public uint Offset11 { get => Data.GetUInt32(_offset11Addr); set => Data.SetUInt32(_offset11Addr, value); }
+        [TableViewModelColumn(addressField: nameof(_offset12Addr), displayOrder: 11, isPointer: true)] [BulkCopy] public uint Offset12 { get => Data.GetUInt32(_offset12Addr); set => Data.SetUInt32(_offset12Addr, value); }
+        [TableViewModelColumn(addressField: nameof(_offset13Addr), displayOrder: 12, isPointer: true)] [BulkCopy] public uint Offset13 { get => Data.GetUInt32(_offset13Addr); set => Data.SetUInt32(_offset13Addr, value); }
+        [TableViewModelColumn(addressField: nameof(_offset14Addr), displayOrder: 13, isPointer: true)] [BulkCopy] public uint Offset14 { get => Data.GetUInt32(_offset14Addr); set => Data.SetUInt32(_offset14Addr, value); }
+        [TableViewModelColumn(addressField: nameof(_offset15Addr), displayOrder: 14, isPointer: true)] [BulkCopy] public uint Offset15 { get => Data.GetUInt32(_offset15Addr); set => Data.SetUInt32(_offset15Addr, value); }
+        [TableViewModelColumn(addressField: nameof(_offset16Addr), displayOrder: 15, isPointer: true)] [BulkCopy] public uint Offset16 { get => Data.GetUInt32(_offset16Addr); set => Data.SetUInt32(_offset16Addr, value); }
     }
 }
